requirements.txt
streamlit==1.28.0
numpy==1.24.3
pillow==10.1.0
ultralytics==8.0.196